+++ runtime/selendra/src/config/mod.rs
@@ -1,9 +1,9 @@
-pub mod consensus_config;
-pub mod dex_config;
-pub mod evm_config;
-pub mod funan_config;
-pub mod governance_config;
-pub mod orml_config;
-pub mod treasury_config;
-pub mod utility_config;
-pub mod bridge_config;
+pub mod consensus;
+pub mod dex;
+pub mod evm;
+pub mod funan;
+pub mod governance;
+pub mod orml;
+pub mod treasury;
+pub mod utility;
+pub mod bridge;
